Compute fixups first and overhaul bootstrap
Previously introduced adjust command is now evaluated when moving
commands from COMMITS to WORKING. For branches, this means that instead
of picking the commits in sequence and _then_ adjusting them with a
patch branch, the fixups are interpolated directly into the original
branch - i.e. for pick c1; pick c2; pick c3; adjust 2 . c4 c5 c6 instead
of picking c1..c3 and then patching them we get pick c1; fixup c4; pick
c2; fixup c5; pick c3; fixup c6 which means that c2 never has to be
rebased without the effects of c4 already applied to c1.

More importantly, this allows a bootstrap commit to be able to peek
ahead to a fixup commit. With this overhaul, it's then possible to
implement the new coreboot command. coreboot is for a full boot cycle -
i.e. requiring the previous commit to be coldstart'd first and then a
patch applied. In this case, when back-porting we need the fixup to be
applied first.