Fix generation of configure on 4.08
The automatic regeneration of configure was often failing on 4.08 with
conflict markers left in configure. Problem is that configure wasn't
being regenerated because of caching.

Simplest solution is to backport 4.09's autogen script to the 4.08
branch which included that tweak. Also has the benefit of removing the
runstatedir Ubuntu stuff, as appropriate.