Add locks branch
stack now records the branch heads used and crafts a tree object
containing a list of those along with all the rr-cache objects needed
for the run. If that tree object does not exist in the locks branch, a
commit is added.

Additionally, stack can take as a parameter the sha of a lock commit and
will use that to obtain the BRANCHES and TARGETS arrays and then
reconstruct just the backport- branches.