Display which rerere items were used in a commit
Actually display the items used and attempt to correlate them with
files. The latter may fail since the hash of the preimage is based on
the conflict markers, not the entire file. Fixing that is non-trivial
(git-hash-object is available, but git-rerere doesn't expose its hashing
function trivially enough). However, it tends to be easy to spot which
file it is, diff it against the final result and manually fix it up.