Fix bootstrap cache lookup with fixups
Empty fixup commits generate an "Empty" entry in the reflog (used to
commit any rerere resolutions) which blocks the caching of Bootstrap
commits which have a blank commit in a fixup patch chain.

Fix it turns out is trivial - just ensure the reflog is only displaying
commits which touched boot/ocamlc in the first place (which is what
we're after).