Change the logic function for is-backport (again)
Simpler observation: although the history is a mess at various points,
the mess is _equivalent_ between the relocatable-base-foo branches and
the individual branches themselves... i.e. a branch is drawn into a
backport not because it is newer than the rebase point, but because it
has the same merge-base _on trunk_ as the rebase point.

Both mechanisms left in place and some limited debugging to ensure they
were producing the same output - the mechanism can still be turned off.