Add a caching mechanism to the backport branches
Similar idea to the relocatable-locks branches, relocatable-cache is now
(ab)used to memoize the construction of a given branch. The process of
spilling COMMITS to WORKING allows a key to be constructed for that
particular rebase, assuming that the same commits are picked and that
the rr-cache files are correct. The commit shas form part of the key and
the rr-files are added as a parent commit of the cache commit.

Thus, having constructed the WORKING set for a given branch, if a
suitable cache is found in relocatable-cache, then an additional commit
contains the updates to the PROVENANCE and ALL_REMAPS maps which allow
the WORKING set already computed to be mapped to the subsequent commits.
Finally, the resulting backport branch is _also_ a parent of the cache.

Slightly WIP in that _new_ locks are not necessarily accurately
constructed (but the cache is correctly _used_ for existing locks),
however it's easy to switch off to do a "clean" run.

Each cache commit which is created is validated (i.e. having created it,
the script then ensures that that cache would have created the same
menu).