Simpler handling for empty fixups
Quieter approach - the commit is checked when the analyse command is
executed, and if the fixup is empty it's dropped. It removes the
look-ahead in the commit function by providing a lookup instead (which
is slightly more principled) and that lookup can be trivially abused to
provide the workaround for legacy empty commits.