Cache "empty" bootstraps
The dynamic-tidying branch features a series of commits which run as
feature1 - bootstrap - feature2 - bootstrap. Part of feature2 is
included in one of the main branches (but is a necessary part of
feature2) which means that when recombined the feature2 commit still
contains changes, but the second bootstrap is unnecessary and, because
of the first bootstrap, results in an empty commit. The empty commit
means that there is no commit touching boot/ocamlc for the scan of the
reflog to find which means that this bootstrap is always performed, only
for the result to be "ignored".

The bootstrap command now detects an empty commit and instead truncates
boot/ocamlc and commits it (for the reflog). The change to the branch is
immediately reverted. In parallel, after merging a cached bootstrap
commit, if boot/ocamlc is missing or zero bytes long, the merge is
reverted.